A plot is revealed!
Literally. We're generating some cool Matlab plots of f(alpha) vs. alpha.
The hard part's going to be understanding and interpreting them. More
research is definitely needed, and I need to be sure that we're actually
plotting f(alpha). Because, from what I understand so far, what we see
indicates a dimension much higher than it should be (around 14). Even worse,
it changes depending on the side length used for the boxes.