Introducing the method of moments.
There's a lot of new material in this update. A one-dimensional reduction, binomialboxcounter, was made to test changes and is currently being used to implement the new method of moments (described in Chhabra & Jensen, 1989). This will then be implemented into the 2-/3-D version. Also included are two visualizations generated in Matlab and the generator program for the binomial cascade (also called a Cantor measure). Histogramming will be removed ASAP, hopefully, because it shouldn't be needed with this new algorithm.