Cleaned up binary/txt forking.
Removed separate versions of programs made for binary/txt files. The current versions can handle anything, as long as the call to the data reading function is made with the file type in mind. A comment is in the code to help you find where it needs to be put in.